+++ historic/src/lib.rs
@@ -130,128 +130,9 @@ What configuration do we need for a client?:
   - AccountId type not necessary I think; only used in legacy RPCs.
 
 */
-
 #![allow(missing_docs)]
 
-pub mod config {
-    use subxt_rpcs::RpcConfig;
-    use scale_type_resolver::TypeResolver;
-    use std::future::Future;
+pub mod config;
+pub mod client;
 
-    pub trait Config {
-        /// The block hash type.
-        type Hash: Hash;
-        /// The shape of our historic type definitions.
-        type HistoricTypes<'a>: TypeResolver where Self: 'a;
-
-        /// Return historic types for a given spec version.
-        fn historic_types_for_spec_version(&'_ self, spec_version: u64) -> Self::HistoricTypes<'_>;
-    }
-
-    /// A trait which is applied to any type that is a valid block hash.
-    pub trait Hash: serde::de::DeserializeOwned + serde::Serialize {}
-    impl<T> Hash for T where T: serde::de::DeserializeOwned + serde::Serialize {}
-
-    #[cfg(feature = "scale-info-legacy")]
-    pub mod substrate {
-        use primitive_types::H256;
-        use scale_info_legacy::{ChainTypeRegistry, TypeRegistrySet};
-        use super::Config;
-
-        pub struct SubstrateConfig {
-            historic_types: ChainTypeRegistry,
-        }
-
-        impl Config for SubstrateConfig {
-            type Hash = H256;
-            type HistoricTypes<'a> = TypeRegistrySet<'a>;
-
-            fn historic_types_for_spec_version(&'_ self, spec_version: u64) -> Self::HistoricTypes<'_> {
-                self.historic_types.for_spec_version(spec_version)
-            }
-        }
-    }
-
-}
-
-pub mod client {
-    use subxt_rpcs::RpcClient;
-    use crate::config::Config;
-
-    pub use online_client::{ OnlineClientT, OnlineClient };
-    pub use offline_client::{ OfflineClientT, OfflineClient };
-
-    mod online_client {
-        use crate::config::Config;
-        use crate::client::OfflineClientT;
-        use subxt_rpcs::RpcClient;
-
-        pub trait OnlineClientT<T: Config>: OfflineClientT<T> {
-            /// Get the RPC client used to communicate with the node.
-            fn rpc_client(&self) -> &RpcClient;
-        }
-    
-        pub struct OnlineClient<T: Config> {
-            /// The configuration for this client.
-            config: T,
-            /// The RPC client used to communicate with the node.
-            rpc_client: RpcClient,
-        }
-
-        impl<T: Config> OnlineClientT<T> for OnlineClient<T> {
-            fn rpc_client(&self) -> &RpcClient {
-                &self.rpc_client
-            }
-        }
-
-        impl<T: Config> OfflineClientT<T> for OnlineClient<T> {
-            fn config(&self) -> &T {
-                &self.config
-            }
-
-            fn spec_version_for_block_number(&self, block_number: u64) -> impl Future<Output = u64> + Send {
-                async move { todo!("Implement this!") }
-            }
-        }
-    }
-
-    mod offline_client {
-        use crate::config::Config;
-
-        pub trait OfflineClientT<T: Config> {
-            /// Get the configuration for this client.
-            fn config(&self) -> &T;
-            /// Get the spec version for some block number.
-            fn spec_version_for_block_number(&self, block_number: u64) -> impl Future<Output = u64> + Send;
-        }
-    
-        pub struct OfflineClient<T: Config> {
-            /// The configuration for this client.
-            config: T,
-        }
-
-        impl <T: Config> OfflineClientT<T> for OfflineClient<T> {
-            fn config(&self) -> &T {
-                &self.config
-            }
-            fn spec_version_for_block_number(&self, block_number: u64) -> impl Future<Output = u64> + Send {
-                async move { todo!("Implement this!") }
-            }
-        }
-    }
-
-    /// A client which is ready to decode data at a specific block number.
-    pub struct ClientAtBlock<'a, T: Config + 'a, Client> {
-        /// The client used to communicate with the node.
-        client: Client,
-        /// Historic types to use at this block number.
-        historic_types: T::HistoricTypes<'a>,
-        /// Metadata to use at this block number.
-        metadata: &'a frame_metadata::RuntimeMetadata
-    }
-}
-
-pub mod extrinsics {
-
-}
